# docs/specs/openapi.json
{"openapi":"3.0.0","info":{"title":"Apollo","version":"0.0.1"},"servers":[{"url":"http://localhost:3000/api","description":"Acesso (DEV)"}],"paths":{"/api/oauth/token":{"post":{"summary":"Gera novo token de acesso a aplicação","tags":["Apollo"],"security":[{"BasicAuth":[]}],"responses":{"200":{"description":"Sucesso na processo de autorização","content":{"application/json":{"schema":{"type":"object","properties":{"token":{"type":"string"}}}}}}}}},"/api/status":{"get":{"summary":"Obtém situação do aplicativo","tags":["Apollo"],"responses":{"200":{"description":"Sucesso na consulta da situação","content":{"application/json":{"schema":{"$ref":"#/components/schemas/BuildDTO"}}}}}}},"/api/vehicles":{"post":{"summary":"Cria um novo veículo","tags":["Apollo"],"security":[{"BearerAuth":[]}],"requestBody":{"description":"Informações-base para cadastro de veículo","content":{"application/json":{"schema":{"$ref":"#/components/schemas/VehiclePostDTO"}}}},"responses":{"200":{"description":"Sucesso no cadastro do veículo","content":{"application/json":{"schema":{"type":"object","properties":{"data":{"$ref":"#/components/schemas/VehicleDTO"}}}}}}}},"get":{"summary":"Obtém listagem de veículos","tags":["Apollo"],"security":[{"BearerAuth":[]}],"responses":{"200":{"description":"Veículo obtido com sucesso"}}}},"/api/vehicles/{id}":{"patch":{"summary":"Atualiza um veículo","tags":["Apollo"],"security":[{"BearerAuth":[]}],"parameters":[{"name":"id","in":"path","description":"ID do veículo","required":true,"schema":{"type":"string"}}],"requestBody":{"description":"Informações-base para atualização de um veículo","content":{"application/json":{"schema":{"$ref":"#/components/schemas/VehiclePutDTO"}}}},"responses":{"200":{"description":"Sucesso na atualização do Veículo","content":{"application/json":{"schema":{"type":"object","properties":{"data":{"$ref":"#/components/schemas/VehicleDTO"}}}}}}}},"get":{"summary":"Atualiza um veículo","tags":["Apollo"],"security":[{"BearerAuth":[]}],"parameters":[{"name":"id","in":"path","description":"ID do veículo","required":true,"schema":{"type":"string"}}],"responses":{"200":{"description":"Veículo obtido com sucesso","content":{"application/json":{"schema":{"type":"object","properties":{"data":{"$ref":"#/components/schemas/VehicleDTO"}}}}}}}},"delete":{"summary":"Remove um veículo","tags":["Apollo"],"security":[{"BearerAuth":[]}],"parameters":[{"name":"id","in":"path","description":"ID do veículo","required":true,"schema":{"type":"string"}}],"responses":{"204":{"description":"Veículo removido com sucesso"}}}}},"components":{"securitySchemes":{"BearerAuth":{"type":"http","scheme":"bearer"},"BasicAuth":{"type":"http","scheme":"basic"}},"schemas":{"BuildDTO":{"type":"object","properties":{"build":{"type":"object","properties":{"date":{"type":"string"},"number":{"type":"number"}}},"git":{"type":"object","properties":{"commit":{"type":"string"}}}}},"VehicleDTO":{"type":"object","properties":{"id":{"type":"string"},"name":{"type":"string"},"model":{"type":"string"},"plate":{"type":"string"},"year":{"type":"number"},"km":{"type":"number"}}},"VehiclePostDTO":{"type":"object","properties":{"name":{"type":"string"},"model":{"type":"string"},"plate":{"type":"string"},"year":{"type":"string"},"km":{"types":"string"}}},"VehiclePutDTO":{"type":"object","properties":{"name":{"type":"string"},"model":{"type":"string"},"plate":{"type":"string"},"year":{"type":"string"},"km":{"types":"string"}}}}},"tags":[]}